add subscriber package
This subscriber will wait as a client for the AES key
coming from the service of the publisher.

This AES key is encrypted by RSA. The received
service answer will be sent to the trusted world for
checking if the key is valid and certified.

This key will be then added in the trusted application
and used for decrypting all incomming messages for
the subscribed topic.