optee_ta_security: add trusted application
This application will run on the secure side. The
secure OS (in this example OPTEE-OS) will decrypt
the source code and run it.

This application contains the necessary operations
to encrypt and decrypt array of bytes and to verify
or certify messaegs using AES / RSA.

The keys are the part of the text section of the
application code. See ta_rsa_keys.h